requirements.txt
Django==1.8
djangorestframework==3.1.1
dj-database-url==0.3.0
gunicorn==19.3.0
Jinja2==2.7.3
MarkupSafe==0.23
psycopg2==2.6

# Opbeat
opbeat==2.1

# Flake8 and dependencies
flake8==2.4.0
mccabe==0.3
pep8==1.5.7
pyflakes==0.8.1

# Tests
py==1.4.26
pytest==2.6.4
pytest-django==2.8.0
